+++ app/src/main/java/com/eelizarraras/workout/core/domine/use_cases/GetActivityUseCase.kt
@@ -6,7 +6,7 @@ import com.eelizarraras.workout.core.domine.repository.DataBaseRepository
 class GetActivityUseCase(
     val repository: DataBaseRepository
 ) {
-    operator fun invoke(uid: Int): ActivityModel? {
+    operator fun invoke(uid: Long): ActivityModel? {
         return repository.getActivity(uid)
     }
 }
